requirements.txt
fastapi
uvicorn
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow
python-multipart
